+++ tests/schemas.py
@@ -80,3 +80,9 @@ class ValidationError(Schema):
 class HTTPError(Schema):
     status_code = String(required=True)
     message = String(required=True)
+
+
+class CustomHTTPError(Schema):
+    status_code = String(required=True)
+    message = String(required=True)
+    custom = String(required=True)